requirements.txt
pyjwt
channels==4.3.1
Django==5.2.6
channels_redis==4.3.0
djangorestframework==3.16.1
psycopg2-binary==2.9.10
django-environ==0.12.0
PyJWT==2.10.1
daphne==4.2.1

